Issue #128: Replace field_information_schedule_{1,2} with taxonomy
After this change: Nodes of type data_set use field_information_schedule to
refer to a term in the information_schedule taxonomy. Those terms have various
values set by references to other taxonomies added in this commit. The values of
these other taxonomies are displayed on the data set along with the name and
number from information_schedule.