Issue #128: Ensure Twig trim is only run on strings
The value for field_semi_active_period is two tokens separated by a space, so it
appears as an empty space when there are no values, so it must have trim. Twig
render ensures that it is always a string. The others do not need trim because
they are single tokens, so when the token is empty the field is empty.